refactor: product 리듀서 -> products 리듀서로 이름 변경
- 기존의 products.ts 의 내용을 유형별로 나눠 저장
- layout 리듀서 적용
